third_party/WebKit/LayoutTests/editing/execCommand/insert_list/insert_list_table_no_hang.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
    '<div contenteditable>',
      '^<table border="1" width="100px"><tr>',
          '<td>foo<br><br><br><span>bar|</span></td>',
      '</tr></table><br>',
    '</div>',
  ],
  'insertOrderedList',
  [
    '<div contenteditable>',
      '<table border="1" width="100px"><tbody><tr>',
        '<td><ol>',
          '<li>^foo</li>',
          '<li><br></li>',
          '<li><br></li>',
          '<li><span>bar|</span></li>',
        '</ol></td>',
      '</tr></tbody></table><br>',
    '</div>',
  ]),
  'InsertOrderedList before TABLE should not hang');
</script>
